+++ test/errors.test.js
@@ -1,0 +1,14 @@
+import loader from '../src';
+
+describe('errors', () => {
+  it('validation errors', () => {
+    const err = () =>
+      loader.call({
+        query: { useRelativePath: 1 },
+        emitFile: true,
+      });
+
+    expect(err).toThrow();
+    expect(err).toThrowErrorMatchingSnapshot();
+  });
+});
