+++ test/name-option.test.js
@@ -1,0 +1,39 @@
+import webpack from './helpers/compiler';
+
+describe('when applied with `name` option', () => {
+  it('matches snapshot for `{String}` value', async () => {
+    const config = {
+      loader: {
+        test: /(png|jpg|svg)/,
+        options: {
+          name: '[hash].[ext]',
+        },
+      },
+    };
+
+    const stats = await webpack('fixture.js', config);
+    const [module] = stats.toJson().modules;
+    const { source } = module;
+
+    expect(source).toMatchSnapshot();
+  });
+
+  it('matches snapshot for `{Function}` value', async () => {
+    const config = {
+      loader: {
+        test: /(png|jpg|svg)/,
+        options: {
+          name() {
+            return '[hash].[ext]';
+          },
+        },
+      },
+    };
+
+    const stats = await webpack('fixture.js', config);
+    const [module] = stats.toJson().modules;
+    const { source } = module;
+
+    expect(source).toMatchSnapshot();
+  });
+});
